SD card "stability" workaround
Workaround that stabilizes the SD card handling: After a reset or a power-on: Wait a while. This is obviously neither a great nor a robust solution, but it increases the amount of readable SD cards greatly. It seems like the more used an SD card gets, the longer the initial startup sequence seems to last.