CRT emulation: First two pipeline steps work
Horizontal and vertical filtering via ascal works. What is still missing: Shadow Mask and Gamma Correction. Moreover: This is a DEBUG version that uses the Audio Processing menu item to toggle between Nearest Neighbor and Sharp Bilinear scaling so that we can decide, which of those should be the standard.