fixed jerky joystick on R3 machines
Since the MAX10 fix, R3 machines had jerky joysticks. The fix is to use the debounced reset signal from the MAX10, which is slightly concerning because this means, that on R3 machines, the reset signal seems to jitter and this jitter did reset the joystick debouncers which lead to the jerky joystick movements.