updated QNICE to dev-d64write
This updates the whole C64 to the very QNICE version that is still WIP but that is able to write to FAT32 formatted SD cards with these constraints: Only write to existing files, only write within the boundaries of exiting files. But this is good enough for writing to D64 files which are always existing and which have a fixed file size.