yellow led: improved user experience
The yellow led that shows a "dirty write cache" now stays on - also during the 2 seconds anti-thrashing pause before everything is finally written to SD card. This creates more clarity for end users (less misunderstandings): just due to the fact that the light stays on, you recognize, that there is still someting going on.