fixed frequency-ratio bug in HDMI-Flicker-Free mode
fixes https://github.com/MJoergen/C64MEGA65/issues/2

It is important that also in the HDMI-Flicker-Free-mode we are using the vanilla clock speed given by CORE_CLK_SPEED_PAL (or CORE_CLK_SPEED_NTSC) and not a speed-adjusted version of this speed. Reason: Otherwise the drift-compensation in generate_drive_ce will compensate for the slower clock speed and ensure an exact 32 MHz frequency even though the system has been slowed down by the HDMI-Flicker-Free. This leads to a different frequency ratio C64 vs 1541 and therefore to incompatibilities such as the one described in the above-mentioned GitHub Issue.