VGA: Improved signal stability
So far, only the RGB signal was phase shifted before providing it to the VDAC (by inverting the VDAC clock). The intension is that the VDAC can sample a nice and steady signal. This commit makes sure that additionally to RGB  also HS and VS are phase shifted and this is done by a small refactoring: VDAC uses now the normal video clock but all the signals are buffered in flip-flops that are triggered on the falling edge of the video clock. Additionally, a PBLOCK was added to ensure that the signal propagation delay from these very flip-flops to the output buffer is as short as possible and as in sync (HS/VS) as possible. Before the PBLOCK the registers were placed all over the FPGA's surface.