Shrank ist of planned features for V5
Due to "nerd time" restrictions in 2023, version 5 will now "only" focus on hard- and software cartridges as well as some smaller features such as the ability to change KERNAL ROMs (e.g. for JiffyDOS). That means: The file VERSIONS.md that contains the plan for Version 5 was reduced and the features that won't make it into Version 5 were added back into ROADMAP.md