CRT: Optimize HyperRAM access speed
This is done by organizing the bank RAM as 4kW instead of 8kB. In other
words, the crt2hyperram module writes 16 bits at a time, instead of only
8 bits. Then when reading it is a matter of choosing between the MSB and
the LSB, based on bit 0 of the CPU address.

In practice, it now takes between 55 and 60 CPU clock cycles to fill up
one 8kB bank.