Fixed handle_hardware_expansion_port
Removed case statement inside the process handle_hardware_expansion_port. The semantics of this process is to handle the hardware side of things, therefore a case statement that is about simulated cartridges is not meant to be in this process. Moreover, it seemed that these "cart_roml_io" and "cart_romh_io" statements were wrong because why should we change these hardware inout ports from high-impedance ('Z') to any value while simulating a software cartridge?