WIP Shell support for cartridge loading
Fixed nasty OPTM_HEAP overflow bug. The symptom was this: "From time to time you might see totally scrambled file names. It can for example happen, when a game or demo asks you to insert the next disk." The reason was, because the heap that contains the linked list of files comes after OPTM_HEAP so the overflow overwrote the linked list of files.