PLA Test Added
Written by Daniël Mantione in April 2023. This program tests the functionality of the Commodore 64 PLA. The program uses hardware on an EasyFlash 1CR cartridge to control the cartridge PLA signals. This means that an EasyFlash 1CR cartridge needs to be inserted before running this program. The cartridge can be in programming mode to boot to BASIC. Flash ROM contents do not matter and won't be modified.