WIP M2M enhancement: Shell auto-loading
Additionally to the ability to manually load ROMs, we are now adding the ability of the Shell to auto-load mandatory or optional ROMs on startup. With this commit the architecture is defined (see globals.vhd) and the Shell is able to get the auto-load data from globals.vhd. The actual load routine is still WIP.