Refactor software-based cartridges
Replaced MiSTer's cartridge.v with our own cartridge.vhd.

The reason is:
In MiSTer, all rom chips are concatenated together and placed
sequentially in SDRAM. Whereas in our core, we have the entire CRT file
in HyperRAM, including the chip headers in between each rom. And some
chips are 16 kB in size, while other chips are 8 kB in size.
I've made various hacks around this (including the "+1" in
crt_cacher.vhd), but the design is fundamentally flawed.

This commit replaces the above with a cleaner design.