Refactored menu.asm and fixed visual menu glitches
The refactoring saves precious QNICE stack space. It fixes the following visual glitches: Help menu items were selected when the user entered the help menu, if the help menu item was "below" a submenu. If the menu item for a virtual drive was within or "below" a submenu, then the following visual glitches occured: When the core was saving to a virtual disk and the user tried to unmount while saving, then the visual mounted indicator in the menu (i.e. "selected") was deleted (i.e. "unselected"). General mount status was wrong, particularly when re-mounting a already mounted disk. If the core was reset while the menu was open, the mounted indicator was not unselected.