Fixes https://github.com/MJoergen/C64MEGA65/issues/58
If you switched the core to JiffyDOS and the core is not able to read the JiffyDOS binaries jd-c64.bin and jd-c1541.bin at a later stage, then the core will not start at all and the screen will stay black. Resolved by temporarily switching back to the standard Kernal in such a situation (maybe missing or wrong SD card, etc.). Next time JiffyDOS can be found, the core will switch back to JiffyDOS. But only, if JiffyDOS was chosen as the default Kernal.