EF3: Deliberately disable Kernal mode and C128 mode
We are deliberately not supporting the Kernal mode x"02" of the EF3, because in Kernal mode, the EF3 manipulates the address line A14. While we could emulate the behavior on our side of the transciever, the problem is, that the transciever and the EF3 would fight" against each other: There might be situations where the core sets A14 to zero while the EF3 sets it to one and this "fight" would lead to quite some current flowing which might damage either the MEGA65's transciever or the EF3's CPLD or other logic parts.