EF3: Fixed reset duration and updated cartridge.md
The original EF3 hardware keeps the core in reset for 7 PHI2 cycles. With this fix, the "kill cartridge" key works in the main menu and we expect in general more stability, e.g. when it comes to freezers. The cartridge documentation contains now detailed usage instructions for the EF3 and the Kung Fu Flash.