Fixed Super Snapshot fastloader by switching the C1541 to DUAL ROM
What happened was: When the JiffyDOS files are found, then the JiffyDOS ROMs are always loaded into the internal ROMs. The C64 already had multiple RAMs/"ROMs" for storing the standard Kernal and the JiffyDOS Kernal, so thet the C64 was OK. But the C1541 had only one RAM/"ROM" active which means that JiffyDOS overwrote the C1541 Standard DOS also in the case when the user selected "Standard" instead of JiffyDOS. Since a C1541 with a JiffyDOS instead of standard DOS behaves nearly like a standard machine, nobody noticed, but this bug was here since Alpha 19 (WIP-V5-A19).