"Auto-enable" the "Simulated Cartridge" switch, when a CRT file has been selected
This convenience function makes sure that the user does not run in non-obvious "strange situations". For example when the use has the hardware cartridge mode on or the REU mode and then chooses to load a *.crt file: Without this enhancement, the auto-reset of the CRT loader would lead to the core going back to BASIC or back to the hardware cartridge that might be plugged in at the moment.

Implements https://github.com/MJoergen/C64MEGA65/issues/44